fix: resolve multiple channel UI and logic bugs (#82)
* fix: resolve multiple channel UI and logic bugs

- Fix Composer send button stuck disabled after attachment upload
- Merge private channels into main Channels list in the sidebar
- Handle WebSocket payload for post deletion real-time updates correctly
- Correct 'Browse Channels' modal join logic with valid `channelService` calls
- Prompt notifications for new channel messages and update unread boldings
- Authorize channel creators to perform updates or deletes on their channels
- Remove redundant casts in `backend/src/api/admin.rs`


* fix(backend): use canonical reaction model in websocket reaction mapping

Switch reaction_added/reaction_removed deserialization from non-existent
crate::models::post::Reaction to canonical crate::models::reaction::Reaction.
Also update create_at/update_at mapping to use the i64 field directly.

---------

+++ backend/src/api/admin.rs
@@ -1559,8 +1559,8 @@ fn check_disk(path: &str) -> DiskHealth {
         };
         if libc::statvfs(c_path.as_ptr(), &mut stat) == 0 {
             let block_size = stat.f_frsize;
-            let total = (stat.f_blocks as u64) * block_size;
-            let available = (stat.f_bavail as u64) * block_size;
+            let total = stat.f_blocks * block_size;
+            let available = stat.f_bavail * block_size;
             let used = total.saturating_sub(available);
             let used_percent = if total > 0 {
                 ((used as f64 / total as f64) * 100.0) as u64
